net: shell: Add allocs command
Move "allocs" command to a separate file.
