+++ subsys/net/lib/shell/CMakeLists.txt
@@ -6,6 +6,8 @@ zephyr_library_include_directories(. ${ZEPHYR_BASE}/subsys/net/lib)
 zephyr_library_include_directories(. ${ZEPHYR_BASE}/subsys/net/ip)
 zephyr_library_link_libraries_ifdef(CONFIG_MBEDTLS mbedTLS)
 
+zephyr_library_sources(allocs.c)
+
 zephyr_library_include_directories(
   ${ZEPHYR_BASE}/kernel/include
   ${ARCH_DIR}/${ARCH}/include
